QR Code detection and pose estimation: Add qr code point sorting. Fix wrong coordinate system estimation.

+++ include/edu_perception/stereo_inference.hpp
@@ -33,14 +33,14 @@ public:
   StereoInference(std::shared_ptr<dai::Device> device, const Parameter parameter);
   virtual ~StereoInference() = default;
 
-  Eigen::Vector3d estimateSpatial(const Eigen::Vector2i coord_a, const Eigen::Vector2i coord_b) const;
+  Eigen::Vector3f estimateSpatial(const Eigen::Vector2i coord_a, const Eigen::Vector2i coord_b) const;
 
 protected:
   float getFocalLengthPixel(const std::size_t pixel_width, const Angle horizontal_fov) const;
   float calculateDepth(const std::size_t disparity_pixel) const;
   int calculateDistance(const Eigen::Vector2i coord_a, const Eigen::Vector2i coord_b) const;
   Angle calculateAngle(const int offset) const;
-  Eigen::Vector3d calculateSpatial(const Eigen::Vector2i coord, const float depth) const;
+  Eigen::Vector3f calculateSpatial(const Eigen::Vector2i coord, const float depth) const;
 
 private:
   const Parameter _parameter;
